+++ payments/src/sdk/sdk.ts
@@ -69,9 +69,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "2.0.4";
-    genVersion = "2.210.6";
-    userAgent = "speakeasy-sdk/typescript 2.0.4 2.210.6 1.0.0 @wingspan/payments";
+    sdkVersion = "2.1.0";
+    genVersion = "2.214.10";
+    userAgent = "speakeasy-sdk/typescript 2.1.0 2.214.10 1.0.0 @wingspan/payments";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);